backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.7
pydantic==2.4.2
pydantic-ai==0.1.0
python-docx2txt==0.8
PyPDF2==3.0.1
reportlab==4.0.4
aiofiles==23.2.1
httpx==0.25.0

